(PiSDF) Separate IsThereALongPathSwitch in a new class file
