+++ plugins/org.preesm.model.pisdf/src/org/preesm/model/pisdf/util/PiSDFMergeabilty.java
@@ -158,7 +158,7 @@ public class PiSDFMergeabilty {
    * @return true if the couple does not introduce cycle
    */
   public static boolean isCycleIntroductionConditionValid(final AbstractActor x, final AbstractActor y) {
-    return !PiSDFTopologyHelper.isMoreThanOneSimplePath(x, y);
+    return !PiSDFTopologyHelper.isThereIsALongPath(x, y);
   }
 
   /**
