+++ plugins/org.preesm.model.pisdf/src/org/preesm/model/pisdf/util/topology/PiSDFSuccessorSwitch.java
@@ -114,15 +114,6 @@ public abstract class PiSDFSuccessorSwitch extends PiMMSwitch<Boolean> {
 
   /**
    *
-   * @author dgageot
-   *
-   */
-  static class MoreThanOneSimplePathException extends RuntimeException {
-    private static final long serialVersionUID = 8123950588521527792L;
-  }
-
-  /**
-   *
    * @author anmorvan
    *
    */
@@ -139,57 +130,6 @@ public abstract class PiSDFSuccessorSwitch extends PiMMSwitch<Boolean> {
         throw new SuccessorFoundException();
       }
       return super.caseAbstractActor(actor);
-    }
-
-  }
-
-  /**
-   *
-   * @author dgageot
-   *
-   */
-  static class IsMoreThanOneSimplePathSwitch extends PiSDFSuccessorSwitch {
-    private final AbstractActor potentialSucc;
-
-    private long fifoCount;
-
-    public IsMoreThanOneSimplePathSwitch(final AbstractActor potentialSucc) {
-      this.potentialSucc = potentialSucc;
-      this.fifoCount = 0;
-    }
-
-    @Override
-    public Boolean caseAbstractActor(final AbstractActor actor) {
-      if (actor.equals(this.potentialSucc)) {
-        if (this.fifoCount > 1) {
-          throw new MoreThanOneSimplePathException();
-        }
-        return true;
-      }
-      if (!this.visitedElements.contains(actor)) {
-        this.visitedElements.add(actor);
-        long tempFifoCount = this.fifoCount;
-        for (DataOutputPort dop : actor.getDataOutputPorts()) {
-          this.fifoCount = tempFifoCount;
-          doSwitch(dop);
-        }
-      }
-      return true;
-    }
-
-    @Override
-    public Boolean caseFifo(final Fifo fifo) {
-      // If there is a delay on fifo, stop here
-      if (fifo.getDelay() != null) {
-        return true;
-      }
-      this.fifoCount = this.fifoCount + 1;
-      return super.caseFifo(fifo);
-    }
-
-    @Override
-    public Boolean casePiGraph(final PiGraph graph) {
-      return caseAbstractActor(graph);
     }
 
   }
